remove ApngFrame struct, fix PngReader method for scaning APNG chunks
It turns out I misunderstood (misremembered?) the docs on APNG frames. The sequence numbers are unique between fcTL and fdAT chunks - they don't define a frame.
And we have to include any IDAT chunks if the fcTL chunk appears before the first IDAT (meaning the first frame of the animation is the 'static image' in the IDAT chunks).